+++ frontend/src/app/shared/interfaces/custom-validators.interface.ts
@@ -1,0 +1,10 @@
+import { DatetimeOption } from "../enums/datetime-options.enum";
+import { DateTimeService } from "../services/datetime.service";
+
+export interface InvalidBHValidatorParams {
+    service: DateTimeService,
+    format: DatetimeOption,
+    startHour: number,
+    endHour: number
+    isPickup: boolean
+}